# Optimize query uniqueness check in ProcessList