Hi!
In our ClickHouse use case, we have a lot of users (currently ~couple thousands per CH instance), all of whom are expected to run queries. 
Recently we tried to push this even further (~couple tens of thousands users per CH instance), which resulted in a significant degradation in our QPS throughput. We were able to achieve only between 100 and 200 QPS, with each query taking up to 30s (expected run time for these queries is under 100ms).

With tracing logs enabled, we were able to figure that most queries were spending their time on this lock:
https://github.com/ClickHouse/ClickHouse/blob/c2f4dc0f14f55cf9785da86cb2bf07c556e22d1e/src/Interpreters/ProcessList.cpp#L82C27-L82C27
The reason was that the inserts into the ProcessList were taking too long, specifically in this for loop:
https://github.com/ClickHouse/ClickHouse/blob/c2f4dc0f14f55cf9785da86cb2bf07c556e22d1e/src/Interpreters/ProcessList.cpp#L185-L194
The for loop iterates over every user the server ever saw, which explains why most queries spend so much time on this lock (we saw times up to 10ms spend in the loop, which would explain the ~100 QPS limit we saw).
This PR, replaces the for loop with a single lookup into a `std::unordered_map`.

We tested this patch in our environment and achieved over 2000 QPS on the same workload as before (could definitely go higher).


### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve performance of executing queries for use cases with many users

### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/