@FrankChen021 2023-12-06
Hi @adikus , I'm wondering in your test, how many users are there? We also notice that the ClickHouse spends time on the ProcessList processing, in our case, there's only 1 user that issues queries up to 2K. I think the 1 user scenario cannot benefit much from this patch, right?