@jsc0218 2023-12-06
> Hi @adikus , I'm wondering in your test, how many users are there? We also notice that the ClickHouse spends time on the ProcessList processing, in our case, there's only 1 user that issues queries up to 2K. I think the 1 user scenario cannot benefit much from this patch, right?

I don't think this patch benefits much in you case, since it mainly avoids the cost of scanning over the user_process_list, while in your case just one user.